renaming GeoRDD to RasterRDD
